Commit do excluir com autenticacao

+++ todolist2/app/Http/Controllers/AskController.php
@@ -8,12 +8,18 @@ use TodoList\Http\Requests;
 
 use TodoList\Ask;
 
+use Auth;
+
 class AskController extends Controller
 {
     function excluir($id){
-      $tarefa = Ask::find($id);
-      $tarefa->delete();
-
-      return redirect('/home')->withInput();
+      if(Auth::check()){
+        $tarefa = Ask::find($id);
+        $tarefa->delete();
+        return redirect('/home')->withInput();
+      }
+      else{
+        return redirect('home');
+      }
     }
 }
